pyproject.toml
[project]
name = "smart-ocr"
version = "0.1.0"
description = "Multi-engine document OCR with cascading fallback"
authors = [
    {name = "Ruben Fernandez-Fuertes", email = "fernandezfuertesruben@gmail.com"}
]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11,<3.13"
keywords = ["ocr", "pdf", "document-processing", "smart-ocr", "nougat", "mistral", "gemini", "deepseek"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "rich>=13.0.0",
    "click>=8.1.0",
    "pillow>=10.0.0",
    "pymupdf>=1.24.0",
    "httpx>=0.27.0",
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.8.0",
]
# OCR Engine CLIs - install individually based on needs
deepseek = ["deepseek-ocr-cli>=0.1.0"]
nougat = ["nougat-ocr-cli>=0.1.2"]
gemini = ["gemini-ocr-cli>=0.2.0"]
mistral = ["mistral-ocr-cli>=0.1.0"]
# Convenience groups
local = ["deepseek-ocr-cli>=0.1.0", "nougat-ocr-cli>=0.1.2"]
cloud = ["gemini-ocr-cli>=0.2.0", "mistral-ocr-cli>=0.1.0"]
all = ["deepseek-ocr-cli>=0.1.0", "nougat-ocr-cli>=0.1.2", "gemini-ocr-cli>=0.2.0", "mistral-ocr-cli>=0.1.0"]

[project.scripts]
smart-ocr = "smart_ocr.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/smart_ocr"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
